Allow per runtime EntryKind
