+++ pallets/data-renewal/src/lib.rs
@@ -32,7 +32,7 @@
 //! ## Cross-pallet contract
 //!
 //! The storage pallet has no renewal vocabulary; this pallet owns all of it through
-//! two opaque payloads the runtime wires ([`EntryKind`] as `EntryMeta`,
+//! two opaque payloads the runtime wires (an `EntryMeta` implementing [`RenewMeta`],
 //! [`PermanentExtent`] as `AuthorizationExtra`):
 //!
 //! - **Down → storage:** dispatchables use the storage pallet's public API; the per-account renew
@@ -67,7 +67,7 @@ mod mock;
 mod tests;
 
 pub use pallet::*;
-pub use types::{EntryKind, PermanentExtent, RenewalData};
+pub use types::{PermanentExtent, RenewMeta, RenewalData};
 pub use weights::WeightInfo;
 
 use bulletin_transaction_storage_primitives::ContentHash;
@@ -117,7 +117,7 @@ pub mod pallet {
 	pub trait Config:
 		frame_system::Config
 		+ pallet_bulletin_transaction_storage::Config<
-			EntryMeta = EntryKind,
+			EntryMeta: RenewMeta,
 			AuthorizationExtra = PermanentExtent,
 		>
 	{
@@ -481,7 +481,7 @@ impl<T: Config> Pallet<T> {
 		let extrinsic_index =
 			<frame_system::Pallet<T>>::extrinsic_index().ok_or(Error::<T>::BadContext)?;
 		pallet_bulletin_transaction_storage::Pallet::<T>::with_block_transactions(|entries| {
-			entries.renew(&info, extrinsic_index, EntryKind::Renew)
+			entries.renew(&info, extrinsic_index, T::EntryMeta::renew())
 		})
 		.ok_or(Error::<T>::TooManyTransactions)
 	}
@@ -528,7 +528,7 @@ impl<T: Config> Pallet<T> {
 				let charged =
 					was_paid || Self::check_renew_authorization(&scope, tx_info.size, true).is_ok();
 				let new_index = if charged {
-					entries.renew(&tx_info, extrinsic_index, EntryKind::Renew)
+					entries.renew(&tx_info, extrinsic_index, T::EntryMeta::renew())
 				} else {
 					None
 				};
@@ -742,7 +742,7 @@ impl<T: Config> Pallet<T> {
 			|acc, (_, entries)| {
 				entries
 					.iter()
-					.filter(|t| t.meta == EntryKind::Renew)
+					.filter(|t| t.meta.is_renew())
 					.fold(acc, |inner, t| inner.saturating_add(t.size as u64))
 			},
 		);
@@ -797,14 +797,14 @@ impl<T: Config> Pallet<T> {
 /// Obsolete-block sweep callback: decrements the chain-wide counter for aged-out
 /// `Renew` entries and queues `is_latest` entries with a [`Renewals`] registration
 /// into [`PendingAutoRenewals`] for the same block's inherent.
-impl<T: Config> OnObsoleteTransactions<BlockNumberFor<T>, EntryKind> for Pallet<T> {
+impl<T: Config> OnObsoleteTransactions<BlockNumberFor<T>, T::EntryMeta> for Pallet<T> {
 	fn handle_obsolete(_obsolete: BlockNumberFor<T>, items: &[(TransactionInfoFor<T>, bool)]) {
 		// Renewed bytes leaving the retention window free up permanent capacity.
 		// Stale shadows (`is_latest == false`) count too: their sizes were charged
 		// when their renew was consumed.
 		let renewed_sum: u64 = items
 			.iter()
-			.filter(|(tx_info, _)| tx_info.meta == EntryKind::Renew)
+			.filter(|(tx_info, _)| tx_info.meta.is_renew())
 			.fold(0u64, |acc, (tx_info, _)| acc.saturating_add(tx_info.size as u64));
 		if renewed_sum > 0 {
 			Self::update_permanent_storage_used(|used| used.saturating_sub(renewed_sum));
@@ -888,7 +888,7 @@ impl<T: Config> txs_benchmarking::BenchmarkHelper<T> for RenewalBenchmarkHelper 
 	}
 
 	fn worst_case_entry_meta() -> T::EntryMeta {
-		EntryKind::Renew
+		T::EntryMeta::renew()
 	}
 
 	fn register_worst_case_entry(content_hash: ContentHash) {
