+++ pallets/data-renewal/src/mock.rs
@@ -40,6 +40,37 @@ type Block = MockBlock<Test>;
 const PRIORITY: TransactionPriority = TransactionPriority::MAX;
 const LONGEVITY: TransactionLongevity = 10;
 
+/// Local stand-in for the runtime-composed `EntryMeta` enum.
+#[derive(
+	Copy,
+	Clone,
+	Debug,
+	PartialEq,
+	Eq,
+	Default,
+	codec::Encode,
+	codec::Decode,
+	codec::MaxEncodedLen,
+	scale_info::TypeInfo,
+)]
+pub enum EntryKind {
+	#[default]
+	#[codec(index = 0)]
+	Store,
+	#[codec(index = 1)]
+	Renew,
+}
+
+impl crate::RenewMeta for EntryKind {
+	fn renew() -> Self {
+		Self::Renew
+	}
+
+	fn is_renew(&self) -> bool {
+		matches!(self, Self::Renew)
+	}
+}
+
 #[frame_support::runtime]
 mod runtime {
 	#[runtime::runtime]
@@ -120,7 +151,7 @@ impl pallet_bulletin_transaction_storage::Config for Test {
 	type RemoveExpiredAccountAuthorizationTxParams = RemoveExpiredAccountAuthorizationTxParams;
 	type RemoveExpiredPreimageAuthorizationTxParams = RemoveExpiredPreimageAuthorizationTxParams;
 	type RemoveExhaustedAuthorizerTxParams = RemoveExhaustedAuthorizerTxParams;
-	type EntryMeta = crate::EntryKind;
+	type EntryMeta = EntryKind;
 	type AuthorizationExtra = crate::PermanentExtent;
 	type OnObsoleteTransactions = DataRenewal;
 	// The runtimes wire this, so the benchmark tests exercise it too.
